feat(#1017): use Representation class to avoid code duplication
