+++ src/test/java/benchmark/DisassembleBenchmark.java
@@ -1,37 +1,14 @@
 /*
- * The MIT License (MIT)
- *
- * Copyright (c) 2016-2025 Objectionary.com
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
+ * SPDX-FileCopyrightText: Copyright (c) 2016-2025 Objectionary.com
+ * SPDX-License-Identifier: MIT
  */
 package benchmark;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Comparator;
-import java.util.Optional;
 import java.util.stream.Stream;
 import org.eolang.jeo.Disassembler;
 import org.openjdk.jmh.annotations.Benchmark;
@@ -83,7 +60,7 @@ public class DisassembleBenchmark {
         final Path input = this.dir.resolve("input");
         this.dir.toFile().deleteOnExit();
         Files.createDirectories(input);
-        Files.write(input.resolve("Integer.class"), this.clazz());
+        Files.write(input.resolve("X.class"), new Representation().bytecode());
         this.disassembler = new Disassembler(input, this.dir);
     }
 
@@ -97,46 +74,7 @@ public class DisassembleBenchmark {
         try (Stream<Path> files = Files.walk(this.dir).sorted(Comparator.reverseOrder())) {
             files.map(Path::toFile).forEach(File::delete);
         } catch (IOException ex) {
-            throw new IllegalStateException("Failed to delete temporary directory", ex);
-        }
-    }
-
-    /**
-     * Disassemble Integer class.
-     * @return Disassembled Integer class.
-     */
-    private byte[] clazz() {
-        final InputStream stream = Thread.currentThread()
-            .getContextClassLoader()
-            .getResourceAsStream(
-                String.format(
-                    "%s.class",
-                    Integer.class.getName().replace('.', '/')
-                )
-            );
-        return this.bytes(
-            Optional.ofNullable(
-                stream
-            ).orElseThrow(() -> new IllegalStateException("Failed to load Integer class"))
-        );
-    }
-
-    /**
-     * Read all bytes from input stream.
-     * @param stream Input stream.
-     * @return Bytes.
-     */
-    private byte[] bytes(final InputStream stream) {
-        try {
-            final ByteArrayOutputStream buffer = new ByteArrayOutputStream();
-            int nread;
-            byte[] data = new byte[stream.available()];
-            while ((nread = stream.read(data, 0, data.length)) != -1) {
-                buffer.write(data, 0, nread);
-            }
-            return buffer.toByteArray();
-        } catch (IOException ex) {
-            throw new IllegalStateException("Failed to read all bytes from input stream", ex);
+            throw new IllegalStateException("Failed to delete disassemble temporary directory", ex);
         }
     }
 }
